--- core-ui/src/shared/components/Categories/Categories.scss
@@ -1,15 +1,0 @@
-.categories-wrapper {
-  display: flex;
-  //   align-content: stretch;
-  flex-wrap: wrap;
-
-  .fd-token {
-    // flex-grow: 1 !important;
-    margin-bottom: 4px;
-    margin-right: 4px;
-    white-space: normal;
-    height: auto;
-    line-height: 1rem;
-    padding: 0.25rem 0.4rem !important;
-  }
-}
