+++ core/src/i18n/en.yaml
@@ -598,8 +598,8 @@ custom-resource-definitions:
     storage: Storage
   subtitle:
     additional-columns: Additional Printer Columns
-    common-crds: Related CRDs
     names: Names
+    related-crds: Related CRDs
     schema: Schema
     version: Version
   title: Custom Resource Definitions
