+++ shared/components/Labels/Labels.scss
@@ -1,10 +1,8 @@
 .labels {
   display: flex;
-  //   align-content: stretch;
   flex-wrap: wrap;
 
   .fd-token {
-    // flex-grow: 1 !important;
     margin-bottom: 4px;
     margin-right: 4px;
     white-space: normal;
